Remove bare-OpenCV methods from DepthEstimator
