+++ lib/src/main/cpp/examples/CMakeLists.txt
@@ -1,5 +1,5 @@
 project("examples")
 
-add_executable(${PROJECT_NAME} get_depth_example.cpp)
+add_executable(${PROJECT_NAME} GetDepthExample.cpp)
 
 target_link_libraries(${PROJECT_NAME} sgbm_android_core)
