+++ lib/src/main/cpp/src/Conversions.h
