+++ lib/src/main/cpp/src/Logging.h
